refactor : type분리, 공용버튼 prop수정
