+++ src/type/AlbumType.ts
@@ -1,0 +1,7 @@
+export type getAlbumContentType = {
+    albumId: number;
+    imageUrl: string;
+    commentCount: number;
+    empathyCount: number;
+    title: string;
+  };
